refactor(desktop): extract model-picker keyboard nav into a hook
ComposerModelSelectorControl crossed the 400-line frontend threshold with
the keyboard-nav addition; the roving-highlight logic now lives in
hooks/chat/ui/use-model-picker-keyboard-nav.ts.
